requirements.txt
discord.py
discord
python-dotenv
colorama~=0.4.4